в метод deletePosition добавлена проверка отсутствия сотрудников с данной позицией.

+++ src/main/java/carservice/service/admin/position/AdminPositionServiceImpl.java
@@ -1,20 +1,29 @@
 package carservice.service.admin.position;
 
+import carservice.exeption.ConflictDataException;
+import carservice.model.employee.EmployeeDtoOut;
 import carservice.model.position.Position;
 import carservice.model.position.PositionDtoIn;
 import carservice.model.position.PositionDtoOut;
 import carservice.model.position.PositionMapping;
 import carservice.repository.PositionRepository;
+import carservice.service.admin.employee.AdminEmployeeService;
 import carservice.service.user.position.UserPositionServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 public class AdminPositionServiceImpl extends UserPositionServiceImpl implements AdminPositionService {
 
+    private final AdminEmployeeService adminEmployeeService;
+
     @Autowired
-    public AdminPositionServiceImpl(PositionRepository positionRepository) {
+    public AdminPositionServiceImpl(PositionRepository positionRepository,
+                                   AdminEmployeeService adminEmployeeService) {
         super(positionRepository);
+        this.adminEmployeeService = adminEmployeeService;
     }
 
     @Override
@@ -34,7 +43,9 @@ public class AdminPositionServiceImpl extends UserPositionServiceImpl implements
     @Override
     public void deletePosition(Long id) {
         getPositionById(id);
-        //TODO реализовать проверку на отсутствие сотрудников с данной должностью
+        List<EmployeeDtoOut> employees = adminEmployeeService.getEmployeesByPosition(id);
+        if(employees.size() > 0) throw new ConflictDataException("Нельзя удалить должность, " +
+                "так как она закреплена за следующими сотрудниками: " + employees);
         positionRepository.deleteById(id);
     }
